requirements.txt
pyTelegramBotAPI==4.6.0
python-dotenv==0.20.0
requests~=2.28.1
python-telegram-bot-calendar==1.0.5
loguru~=0.6.0
